requirements.txt
slack-bolt>=1.18,<2
slack-sdk>=3.27,<4
httpx>=0.27,<1
python-dotenv>=1.0,<2
